Semi-implemented native objects (more in description)
I'm not sure if this is good in terms of backwards compatibility, but I'll hope that I won't need to change much later on.

It is "semi-implemented" because not all objects are supported. For example, types are a Python type which should be natively converted to their matching counterparts for C++. A case where this could cause an error if not done is in function arguments.

See this commit's "testcode.cpp", the annotation "str" was moved directly to the transpilation instead of being converted to std::string and requiring the relevant dependency.